requirements.txt
imageio
scipy
six
numpy
einops
pillow
python-dateutil==2.7.3
torch==1.7.1
torchvision==0.8.2
tensorboard==1.12.2
tensorboardX==1.6
tensorflow==2.4
tqdm==4.29.1
opencv-python
